refactor: Refactored AccountApplication Contracts Divided Admin & User
