+++ AccountManagement.Application.Contracts/Account/Admin/RegisterAccount.cs
@@ -3,7 +3,7 @@ using _0_Framework.Application;
 using AccountManagement.Application.Contracts.Role;
 using Microsoft.AspNetCore.Http;
 
-namespace AccountManagement.Application.Contracts.Account
+namespace AccountManagement.Application.Contracts.Account.Admin
 {
     public class RegisterAccount
     {
@@ -19,12 +19,12 @@ namespace AccountManagement.Application.Contracts.Account
         [MaxLength(200, ErrorMessage = " {0} نمیتواند بیشتر از {1} کاراکترباشد")]
         [EmailAddress(ErrorMessage = ValidationMessages.EmailIsNotValid)]
         public string Email { get; set; }
-        
+
         [Required(ErrorMessage = ValidationMessages.IsRequired)]
         [MaxLength(11, ErrorMessage = " {0} نمیتواند بیشتر از {1} کاراکترباشد")]
         public string Mobile { get; set; }
-        
-        [Range(1, int.MaxValue , ErrorMessage = ValidationMessages.IsRequired)]
+
+        [Range(1, int.MaxValue, ErrorMessage = ValidationMessages.IsRequired)]
         public int RoleId { get; set; }
 
         [Required(ErrorMessage = ValidationMessages.IsRequired)]
