+++ AccountManagement.Application.Contracts/Account/IAccountApplication.cs
@@ -1,4 +1,6 @@
 ﻿using _0_Framework.Application;
+using AccountManagement.Application.Contracts.Account.Admin;
+using AccountManagement.Application.Contracts.Account.User;
 
 namespace AccountManagement.Application.Contracts.Account
 {
