+++ AccountManagement.Application.Contracts/Account/User/AccountInformationViewModel.cs
@@ -1,13 +1,13 @@
-﻿namespace AccountManagement.Application.Contracts.Account
+﻿namespace AccountManagement.Application.Contracts.Account.User
 {
     public class AccountInformationViewModel
     {
         public string Username { get; set; }
-        
+
         public string Email { get; set; }
-        
+
         public string Mobile { get; set; }
-        
+
         public string ProfilePhoto { get; set; }
 
         public string RegisterDate { get; set; }
