+++ AccountManagement.Application.Contracts/Account/User/Login.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using _0_Framework.Application;
 
-namespace AccountManagement.Application.Contracts.Account
+namespace AccountManagement.Application.Contracts.Account.User
 {
     public class Login
     {
