+++ AccountManagement.Application.Contracts/Account/User/RegisterAccountByUser.cs
@@ -1,7 +1,8 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using _0_Framework.Application;
+using AccountManagement.Application.Contracts.Account.Admin;
 
-namespace AccountManagement.Application.Contracts.Account
+namespace AccountManagement.Application.Contracts.Account.User
 {
     public class RegisterAccountByUser : RegisterAccount
     {
