+++ AccountManagement.Application.Contracts/Account/User/ResetPasswordViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace AccountManagement.Application.Contracts.Account
+namespace AccountManagement.Application.Contracts.Account.User
 {
     public class ResetPasswordViewModel
     {
