+++ ServiceHost/Areas/Administration/Pages/Accounts/Role/Index.cshtml.cs
@@ -1,4 +1,4 @@
-using AccountManagement.Application.Contracts.Account;
+using AccountManagement.Application.Contracts.Account.Admin;
 using AccountManagement.Application.Contracts.Role;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
