+++ ServiceHost/Areas/Administration/Pages/Accounts/User/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model AccountManagement.Application.Contracts.Account.RegisterAccount
+﻿@model AccountManagement.Application.Contracts.Account.Admin.RegisterAccount
 @{
 }
