+++ ServiceHost/Areas/Administration/Pages/Accounts/User/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model AccountManagement.Application.Contracts.Account.EditAccount
+﻿@model AccountManagement.Application.Contracts.Account.Admin.EditAccount
 @{
 }
