+++ ServiceHost/Pages/Shared/Components/UserSideBar/Default.cshtml
@@ -1,4 +1,5 @@
 ﻿@using AccountManagement.Application.Contracts.Account
+@using AccountManagement.Application.Contracts.Account.User
 @inject IAccountApplication _accountApplication
 
 @{
