+++ ServiceHost/Pages/Shared/_SuccessRegister.cshtml
@@ -1,4 +1,4 @@
-﻿@model AccountManagement.Application.Contracts.Account.RegisterAccountByUser
+﻿@model AccountManagement.Application.Contracts.Account.User.RegisterAccountByUser
 
 @{
     ViewData["Title"] = "ثبت نام با موفقیت انجام شد";
